#2091: splitting off and renaming DynKernCallFactory
